requirements.txt
aiohttp==3.7.2
aioredis==1.3.1
appnope==0.1.0
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
bottle==0.12.19
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
decorator==4.4.2
ecdsa==0.14.1
fastapi==0.61.2
Flask==1.1.2
h11==0.11.0
hiredis==1.1.0
idna==2.10
iniconfig==1.1.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.0.2
packaging==20.4
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.2
Pygments==2.7.2
pyparsing==2.4.7
pytest==6.1.2
python-jose==3.2.0
requests==2.25.0
rsa==4.6
six==1.15.0
starlette==0.13.6
toml==0.10.2
traitlets==5.0.5
typing-extensions==3.7.4.3
ujson==4.0.1
urllib3==1.26.2
uvicorn==0.12.2
uWSGI==2.0.19.1
wcwidth==0.2.5
Werkzeug==1.0.1
yarl==1.6.3
